IUserContextService

+++ RestaurantAPI_v2/Services/RestaurantService.cs
@@ -1,12 +1,15 @@
 ﻿using AutoMapper;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Logging;
+using RestaurantAPI_v2.Authorization;
 using RestaurantAPI_v2.Entities;
 using RestaurantAPI_v2.Exceptions;
 using RestaurantAPI_v2.Models;
 using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Security.Claims;
 using System.Text;
 using System.Threading.Tasks;
 
@@ -25,20 +28,35 @@ namespace RestaurantAPI_v2.Services
         private readonly RestaurantDbContext _dbContext;
         private readonly IMapper _mapper;
         private readonly ILogger<RestaurantService> _logger;
-        public RestaurantService(RestaurantDbContext dbContext, IMapper mapper, ILogger<RestaurantService> logger)
+        private readonly IAuthorizationService _authorizationService;
+        private readonly IUserContextService _userContextService;
+
+        public RestaurantService(RestaurantDbContext dbContext, IMapper mapper, ILogger<RestaurantService> logger
+            , IAuthorizationService authorizationService, IUserContextService userContextService)
         {
             _mapper = mapper;
             _dbContext = dbContext;
             _logger = logger;
+            _authorizationService = authorizationService;
+            _userContextService = userContextService;
         }
         public void ModifyRestaurant(int id, ModifyRestaurantDto dto)
         {
+
             var restaurant = _dbContext
                 .Restaurants
                 .FirstOrDefault(x => x.Id == id);
 
             if (restaurant is null)
                 throw new NotFoundException("Restaurant not found");
+
+            var authorisationResult =  _authorizationService.AuthorizeAsync(_userContextService.User, restaurant,
+                new ResourceOperationRequirement(ResourceOperation.Update)).Result;
+
+            if (!authorisationResult.Succeeded)
+            {
+                throw new ForbidException();
+            }
 
             restaurant.Name = dto.Name;
             restaurant.Description = dto.Description;
@@ -60,6 +78,14 @@ namespace RestaurantAPI_v2.Services
 
             if (restaurant == null)
                 throw new NotFoundException("Restaurant not found");
+
+            var authorisationResult = _authorizationService.AuthorizeAsync(_userContextService.User, restaurant,
+                new ResourceOperationRequirement(ResourceOperation.Delete)).Result;
+
+            if (!authorisationResult.Succeeded)
+            {
+                throw new ForbidException();
+            }
 
             _dbContext.Restaurants.Remove(restaurant);
             _dbContext.SaveChanges();
@@ -98,6 +124,7 @@ namespace RestaurantAPI_v2.Services
         public int Create(CreateRestaurantDto dto)
         {
             var restaurant = _mapper.Map<Restaurant>(dto);
+            restaurant.CreatedById = _userContextService.GetUserId;
             _dbContext.Restaurants.Add(restaurant);
             _dbContext.SaveChanges();
